+++ CHANGES.md
@@ -3,6 +3,7 @@
 * load in `__construct()`
 * update error email message
 * don't check our own plugin, cause it will always have a PHP fatal for redeclaring the class
+* refactor for restarting plugin update process after a fatal
 
 #### 0.6.0 / 2022-08-05
 * update documentation