+++ CHANGES.md
@@ -4,6 +4,7 @@
 * update error email message
 * don't check our own plugin, cause it will always have a PHP fatal for redeclaring the class
 * refactor for restarting plugin update process after a fatal
+* add function for sending successful update email
 
 #### 0.6.0 / 2022-08-05
 * update documentation